cloud_function_1/requirements.txt
google-cloud-bigquery
google-cloud-storage
flask
functions-framework
